Updating TPC modules to replace floats with doubles.

+++ offline/packages/tpc/LaserEventInfov1.cc
@@ -20,7 +20,7 @@ void LaserEventInfov1::Reset()
   for (int i = 0; i < 2; i++)
   {
     m_peakSample[i] = std::numeric_limits<int>::max();
-    m_peakWidth[i] = std::numeric_limits<float>::quiet_NaN();
+    m_peakWidth[i] = std::numeric_limits<double>::quiet_NaN();
   }
 
   return;
